fix: consolidate de-part libs to satisfy app main line budget (#4117)
Merge small explicit-import libraries and trim comments so
repo.app_main_line_budget passes after wave-9 de-part import overhead.
Unblocks PR #4128 Quality gate failure.
