+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel.dart
@@ -6,5 +6,5 @@ export 'diplomacy_panel_chrome_standing.dart'
     show DiplomacyStandingChipCluster;
 export 'diplomacy_panel_constants.dart';
 export 'diplomacy_panel_rows.dart';
-export 'diplomacy_panel_widget.dart';
+export 'diplomacy_panel_state.dart' show DiplomacyPanel;
 export 'relative_power_line.dart';
