+++ app/lib/widgets/main_menu_constants.dart
@@ -105,3 +105,14 @@ const double kMainMenuScrollBracketVerticalInset = 0.10;
 /// dots). Mirrors mockup `.buttons-region::before { opacity: 0.45 }` so the
 /// bracket reads as a faint chrome flourish, not a structural rule.
 const double kMainMenuScrollBracketOpacity = 0.45;
+
+enum MainMenuVariant {
+  plain,
+  pixelArt,
+}
+
+enum MainMenuState {
+  default_,
+  afterVictory,
+  noSaves,
+}
