+++ app/test/widgets/diplomacy_panel_modular_split_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel_body.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel_constants.dart';
-import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel_widget.dart';
+import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel_state.dart';
 
 /// De-parted diplomacy-panel library seam checks (Refs #4117).
 void main() {
